# specs/master/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Physical AI Textbook API
  description: API for AI-native textbook platform with RAG chatbot, personalization, and Urdu translation
  version: 1.0.0
  contact:
    name: API Support
    email: support@physicalai-book.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.physicalai-book.com
    description: Production server
...
paths:
...
  /api/chat:
    post:
      tags:
        - Chat
      summary: Send message and get AI response
      operationId: sendChatMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: AI response with sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded
        '500':
          description: Server error
...
  /api/translate/urdu:
    post:
      tags:
        - Translation
      summary: Translate chapter content to Urdu
      operationId: translateToUrdu
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateRequest'
      responses:
        '200':
          description: Translation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Translation failed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from BetterAuth session
  schemas:
...
    ChatRequest:
      type: object
      properties:
        chapterId:
          type: string
          format: uuid
        message:
          type: string
          minLength: 1
          maxLength: 2000
        conversationHistory:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
          maxItems: 10
      required:
        - chapterId
        - message
    ChatResponse:
      type: object
      properties:
        success:
          type: boolean
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
    Source:
      type: object
      properties:
        chapterId:
          type: string
        section:
          type: string
        score:
          type: number
          format: float

    # ==================== Personalization Schemas ====================
...
    TranslateRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        language:
          type: string
          enum: [urdu]
          default: urdu
      required:
        - content
    TranslateResponse:
      type: object
      properties:
        translated:
          type: string
        fromCache:
          type: boolean
        language:
          type: string
          enum: [urdu]